+++ src/backends/notmuch/notmuch_envelope.rs
@@ -36,10 +36,10 @@ impl DerefMut for NotmuchEnvelopes {
 }
 
 impl PrintTable for NotmuchEnvelopes {
-    fn print_table(&self, writter: &mut dyn WriteColor, opts: PrintTableOpts) -> Result<()> {
-        writeln!(writter)?;
-        Table::print(writter, self, opts)?;
-        writeln!(writter)?;
+    fn print_table(&self, writer: &mut dyn WriteColor, opts: PrintTableOpts) -> Result<()> {
+        writeln!(writer)?;
+        Table::print(writer, self, opts)?;
+        writeln!(writer)?;
         Ok(())
     }
 }