+++ src/output/print_table.rs
@@ -9,7 +9,7 @@ pub trait WriteColor: io::Write + termcolor::WriteColor {}
 impl WriteColor for StandardStream {}
 
 pub trait PrintTable {
-    fn print_table(&self, writter: &mut dyn WriteColor, opts: PrintTableOpts) -> Result<()>;
+    fn print_table(&self, writer: &mut dyn WriteColor, opts: PrintTableOpts) -> Result<()>;
 }
 
 pub struct PrintTableOpts<'a> {